refactor : 인터페이스 의존하도록 변경
