+++ src/main/java/com/fires/fires/security/filter/JwtAuthenticationFilter.java
@@ -8,8 +8,9 @@ import com.fires.fires.common.exception.ResourceNotFoundException;
 import com.fires.fires.member.constant.Provider;
 import com.fires.fires.member.controller.MemberController;
 import com.fires.fires.member.dto.MemberDto;
-import com.fires.fires.member.dto.MemberSignupRequest;
+import com.fires.fires.member.dto.request.MemberSignupRequest;
 import com.fires.fires.member.service.MemberReadService;
+import com.fires.fires.member.service.MemberReadServiceImpl;
 import com.fires.fires.security.dto.UserPrincipal;
 import com.fires.fires.security.util.JwtTokenUtil;
 import jakarta.servlet.FilterChain;
